imported HtmlUnit 3.11.0

+++ src/main/java/org/htmlunit/HttpWebConnection.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2002-2023 Gargoyle Software Inc.
+ * Copyright (c) 2002-2024 Gargoyle Software Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -62,6 +62,7 @@ import org.apache.http.auth.Credentials;
 import org.apache.http.client.AuthCache;
 import org.apache.http.client.CredentialsProvider;
 import org.apache.http.client.config.RequestConfig;
+import org.apache.http.client.methods.CloseableHttpResponse;
 import org.apache.http.client.methods.HttpDelete;
 import org.apache.http.client.methods.HttpGet;
 import org.apache.http.client.methods.HttpHead;
@@ -146,7 +147,7 @@ public class HttpWebConnection implements WebConnection {
     private final WebClient webClient_;
 
     private String virtualHost_;
-    private final CookieSpecProvider htmlUnitCookieSpecProvider_;
+    private final HtmlUnitCookieSpecProvider htmlUnitCookieSpecProvider_;
     private final WebClientOptions usedOptions_;
     private PoolingHttpClientConnectionManager connectionManager_;
 
@@ -188,10 +189,12 @@ public class HttpWebConnection implements WebConnection {
             final long startTime = System.currentTimeMillis();
 
             final HttpContext httpContext = getHttpContext();
-            HttpResponse httpResponse;
             try {
                 try (CloseableHttpClient closeableHttpClient = builder.build()) {
-                    httpResponse = closeableHttpClient.execute(httpHost, httpMethod, httpContext);
+                    try (CloseableHttpResponse httpResponse =
+                            closeableHttpClient.execute(httpHost, httpMethod, httpContext)) {
+                        return downloadResponse(httpMethod, webRequest, httpResponse, startTime);
+                    }
                 }
             }
             catch (final SSLPeerUnverifiedException s) {
@@ -199,12 +202,13 @@ public class HttpWebConnection implements WebConnection {
                 if (webClient_.getOptions().isUseInsecureSSL()) {
                     HtmlUnitSSLConnectionSocketFactory.setUseSSL3Only(httpContext, true);
                     try (CloseableHttpClient closeableHttpClient = builder.build()) {
-                        httpResponse = closeableHttpClient.execute(httpHost, httpMethod, httpContext);
+                        try (CloseableHttpResponse httpResponse =
+                                closeableHttpClient.execute(httpHost, httpMethod, httpContext)) {
+                            return downloadResponse(httpMethod, webRequest, httpResponse, startTime);
+                        }
                     }
                 }
-                else {
-                    throw s;
-                }
+                throw s;
             }
             catch (final Error e) {
                 // in case a StackOverflowError occurs while the connection is leased, it won't get released.
@@ -214,10 +218,6 @@ public class HttpWebConnection implements WebConnection {
                 httpClientBuilder_.remove(Thread.currentThread());
                 throw e;
             }
-
-            final DownloadedContent downloadedBody = downloadResponseBody(httpResponse);
-            final long endTime = System.currentTimeMillis();
-            return makeWebResponse(httpResponse, webRequest, downloadedBody, endTime - startTime);
         }
         finally {
             if (httpMethod != null) {
@@ -677,7 +677,8 @@ public class HttpWebConnection implements WebConnection {
         builder.setSSLSocketFactory(socketFactory);
 
         usedOptions_.setUseInsecureSSL(options.isUseInsecureSSL());
-        usedOptions_.setSSLClientCertificateStore(options.getSSLClientCertificateStore());
+        usedOptions_.setSSLClientCertificateKeyStore(options.getSSLClientCertificateStore(),
+                        options.getSSLClientCertificatePassword());
         usedOptions_.setSSLTrustStore(options.getSSLTrustStore());
         usedOptions_.setSSLClientCipherSuites(options.getSSLClientCipherSuites());
         usedOptions_.setSSLClientProtocols(options.getSSLClientProtocols());
@@ -722,9 +723,14 @@ public class HttpWebConnection implements WebConnection {
     }
 
     /**
-     * Converts an HttpMethod into a WebResponse.
+     * Converts an HttpMethod into a {@link WebResponse}.
+     * @param httpResponse the web server's response
+     * @param webRequest the {@link WebRequest}
+     * @param responseBody the {@link DownloadedContent}
+     * @param loadTime the download time
+     * @return a wrapper for the downloaded body.
      */
-    private WebResponse makeWebResponse(final HttpResponse httpResponse,
+    protected WebResponse makeWebResponse(final HttpResponse httpResponse,
             final WebRequest webRequest, final DownloadedContent responseBody, final long loadTime) {
 
         String statusMessage = httpResponse.getStatusLine().getReasonPhrase();
@@ -741,6 +747,26 @@ public class HttpWebConnection implements WebConnection {
     }
 
     /**
+     * Downloads the response.
+     * This calls {@link #downloadResponseBody(HttpResponse)} and constructs the {@link WebResponse}.
+     * @param httpMethod the HttpUriRequest
+     * @param webRequest the {@link WebRequest}
+     * @param httpResponse the web server's response
+     * @param startTime the download start time
+     * @return a wrapper for the downloaded body.
+     * @throws IOException in case of problem reading/saving the body
+     */
+    protected WebResponse downloadResponse(final HttpUriRequest httpMethod,
+            final WebRequest webRequest, final HttpResponse httpResponse,
+            final long startTime) throws IOException {
+
+        final DownloadedContent downloadedBody = downloadResponseBody(httpResponse);
+        final long endTime = System.currentTimeMillis();
+
+        return makeWebResponse(httpResponse, webRequest, downloadedBody, endTime - startTime);
+    }
+
+    /**
      * Downloads the response body.
      * @param httpResponse the web server's response
      * @return a wrapper for the downloaded body.
@@ -753,7 +779,8 @@ public class HttpWebConnection implements WebConnection {
         }
 
         try (InputStream is = httpEntity.getContent()) {
-            return downloadContent(is, webClient_.getOptions().getMaxInMemory());
+            return downloadContent(is, webClient_.getOptions().getMaxInMemory(),
+                        webClient_.getOptions().getTempFileDirectory());
         }
     }
 
@@ -761,10 +788,12 @@ public class HttpWebConnection implements WebConnection {
      * Reads the content of the stream and saves it in memory or on the file system.
      * @param is the stream to read
      * @param maxInMemory the maximumBytes to store in memory, after which save to a local file
+     * @param tempFileDirectory the directory to be used or null for the system default
      * @return a wrapper around the downloaded content
      * @throws IOException in case of read issues
      */
-    public static DownloadedContent downloadContent(final InputStream is, final int maxInMemory) throws IOException {
+    public static DownloadedContent downloadContent(final InputStream is, final int maxInMemory,
+            final File tempFileDirectory) throws IOException {
         if (is == null) {
             return new DownloadedContent.InMemory(null);
         }
@@ -775,9 +804,9 @@ public class HttpWebConnection implements WebConnection {
             try {
                 while ((nbRead = is.read(buffer)) != -1) {
                     bos.write(buffer, 0, nbRead);
-                    if (bos.size() > maxInMemory) {
+                    if (maxInMemory > 0 && bos.size() > maxInMemory) {
                         // we have exceeded the max for memory, let's write everything to a temporary file
-                        final File file = File.createTempFile("htmlunit", ".tmp");
+                        final File file = File.createTempFile("htmlunit", ".tmp", tempFileDirectory);
                         file.deleteOnExit();
                         try (OutputStream fos = Files.newOutputStream(file.toPath())) {
                             bos.writeTo(fos); // what we have already read
